+++ packages/orq-rc/RELEASES.md
@@ -4599,3 +4599,13 @@ Based on:
 - [typescript v3.14.0-rc.60] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.60] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.60 - packages/orq-rc
+
+## 2025-10-27 11:30:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.639.3 (2.730.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.61] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.61] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.61 - packages/orq-rc